Fixed UI bug, button not visible, refactored code and added test files to repo
