+++ backend/package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "start": "node dist/app.js",
     "dev": "ts-node-dev src/app.ts",
+    "test": "jest",
     "build": "tsc"
   },
   "author": "",
@@ -30,11 +31,16 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/ioredis": "^4.28.10",
+    "@types/jest": "^29.5.14",
     "@types/multer": "^1.4.7",
     "@types/node": "^18.19.67",
     "@types/pdf-parse": "^1.1.4",
     "@types/redis": "^4.0.10",
+    "@types/supertest": "^6.0.2",
+    "jest": "^29.7.0",
     "nodemon": "^3.1.7",
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.5",
     "tsx": "^4.19.1",
     "typescript": "^5.3.1"
   }
