+++ backend/tsconfig.json
@@ -7,7 +7,8 @@
       "strict": true,
       "esModuleInterop": true,
       "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true
+      "forceConsistentCasingInFileNames": true,
+      "types": ["node", "jest"],
     },
     "include": ["src/**/*"]
   }
